requirements.txt
pyotp==2.6.0
requests==2.22.0
setproctitle==1.2.2
jsonpath2==0.4.4
numpy==1.18.1
falcon==3.0.0a3
tensorboardX==2.0
empyrical==0.5.5
scipy==1.7.1
gunicorn==20.1.0
matplotlib==3.1.2
diskcache==5.2.1
gym==0.17.0
zstd==1.4.5.1
dill==0.3.2
pyaxo-ng==1.0.1
beautifulsoup4==4.9.1
lxml==4.5.2
psutil==5.7.0
spectree==0.4.2
pydantic==1.7.3
stripe==2.55.0
sentry-sdk[falcon]==0.16.2
pydantic==1.7.3
result==0.5.0
sendgrid==6.6.0
portalocker==2.3.0
pandas==1.2.5
python-telegram-bot==13.7
tweepy==4.0.0
typing_extensions
newrelic
google-api-core==2.1.0
google-api-python-client==2.25.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
web3==5.24.0
git+https://github.com/i404788/pyti.git@master
git+https://github.com/lwcolton/falcon-cors@fix-static-route
